[#75] refactor: ChatMessage Repository 가 아닌 Query Port 로 변경
