+++ live-chat-server/src/main/java/org/giglab/live/application/port/persistence/ChatMessageQueryPort.java
@@ -1,11 +1,9 @@
-package org.giglab.live.domain.repository;
+package org.giglab.live.application.port.persistence;
 
 import java.util.List;
 import org.giglab.live.domain.model.ChatMessage;
 
-public interface ChatMessageRepository {
-
-  void save(ChatMessage message);
+public interface ChatMessageQueryPort {
 
   List<ChatMessage> findRecentByRoomId(String roomId, int limit);
 }
