+++ live-chat-server/src/main/java/org/giglab/live/application/port/persistence/ChatMessageStorePort.java
@@ -1,0 +1,8 @@
+package org.giglab.live.application.port.persistence;
+
+import org.giglab.live.domain.model.ChatMessage;
+
+public interface ChatMessageStorePort {
+
+  void save(ChatMessage chatMessage);
+}
