+++ live-chat-server/src/main/java/org/giglab/live/application/port/persistence/RoomQueryPort.java
@@ -1,0 +1,3 @@
+package org.giglab.live.application.port.persistence;
+
+public interface RoomQueryPort {}
