+++ live-chat-server/src/main/java/org/giglab/live/infrastructure/mongo/MongoChatMessageRepository.java
@@ -3,8 +3,9 @@ package org.giglab.live.infrastructure.mongo;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.giglab.live.application.port.persistence.ChatMessageQueryPort;
+import org.giglab.live.application.port.persistence.ChatMessageStorePort;
 import org.giglab.live.domain.model.ChatMessage;
-import org.giglab.live.domain.repository.ChatMessageRepository;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
@@ -14,7 +15,7 @@ import org.springframework.stereotype.Repository;
 @Slf4j
 @Repository
 @RequiredArgsConstructor
-public class MongoChatMessageRepository implements ChatMessageRepository {
+public class MongoChatMessageRepository implements ChatMessageStorePort, ChatMessageQueryPort {
 
   private final MongoTemplate mongoTemplate;
